+++ x/auth/testutil/app_config.go
@@ -1,7 +1,6 @@
 package testutil
 
 import (
-	_ "cosmossdk.io/x/feegrant/module"                // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/auth"           // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config" // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/auth/vesting"   // import as blank for app wiring
@@ -12,7 +11,6 @@ import (
 	_ "github.com/cosmos/cosmos-sdk/x/staking"        // import as blank for app wiring
 
 	"cosmossdk.io/core/appconfig"
-	"cosmossdk.io/x/feegrant"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	vestingtypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -27,7 +25,6 @@ import (
 	authmodulev1 "cosmossdk.io/api/cosmos/auth/module/v1"
 	bankmodulev1 "cosmossdk.io/api/cosmos/bank/module/v1"
 	consensusmodulev1 "cosmossdk.io/api/cosmos/consensus/module/v1"
-	feegrantmodulev1 "cosmossdk.io/api/cosmos/feegrant/module/v1"
 	genutilmodulev1 "cosmossdk.io/api/cosmos/genutil/module/v1"
 	paramsmodulev1 "cosmossdk.io/api/cosmos/params/module/v1"
 	stakingmodulev1 "cosmossdk.io/api/cosmos/staking/module/v1"
@@ -44,19 +41,16 @@ var AppConfig = appconfig.Compose(&appv1alpha1.Config{
 				BeginBlockers: []string{
 					stakingtypes.ModuleName,
 					genutiltypes.ModuleName,
-					feegrant.ModuleName,
 				},
 				EndBlockers: []string{
 					stakingtypes.ModuleName,
 					genutiltypes.ModuleName,
-					feegrant.ModuleName,
 				},
 				InitGenesis: []string{
 					authtypes.ModuleName,
 					banktypes.ModuleName,
 					stakingtypes.ModuleName,
 					genutiltypes.ModuleName,
-					feegrant.ModuleName,
 					paramstypes.ModuleName,
 					consensustypes.ModuleName,
 					vestingtypes.ModuleName,
@@ -104,11 +98,6 @@ var AppConfig = appconfig.Compose(&appv1alpha1.Config{
 		{
 			Name:   genutiltypes.ModuleName,
 			Config: appconfig.WrapAny(&genutilmodulev1.Module{}),
-		},
-
-		{
-			Name:   feegrant.ModuleName,
-			Config: appconfig.WrapAny(&feegrantmodulev1.Module{}),
 		},
 	},
 })